+++ CHANGELOG.md
@@ -8,10 +8,11 @@ All notable changes to this project will be documented in this file.
 
 - Deploy default and support custom affinities ([#319]).
 - Added OLM bundle files ([#328]).
+- Extend cluster resources for status and cluster operation (paused, stopped) ([#337]).
 
 ### Changed
 
-- `operator-rs` `0.36.0` -> `0.37.0` ([#326]).
+- `operator-rs` `0.36.0` -> `0.39.0` ([#326], [#337]).
 - [Breaking] Moved top level config option to `clusterConfig` ([#326]).
 
 ### Removed
@@ -22,6 +23,7 @@ All notable changes to this project will be documented in this file.
 [#326]: https://github.com/stackabletech/hdfs-operator/pull/326
 [#328]: https://github.com/stackabletech/hdfs-operator/pull/328
 [#332]: https://github.com/stackabletech/hdfs-operator/pull/332
+[#337]: https://github.com/stackabletech/hdfs-operator/pull/337
 
 ## [23.1.0] - 2023-01-23